Is this an empty file? I think not! You slyly say "commit aborted," And have no thought as to what it's got.
This is not an empty file,
but rather a collection of strings,
and it may go on a while.

Why must you be so difficult to use?
The pros say with ease,
"It's not a ruse!"

Alas my rant must come to end.
Git, you're learning curve may not be steep,
but kiss my rear end.